+++ waldreg/application-layer/board-controller/src/main/java/org/waldreg/controller/board/board/BoardController.java
@@ -17,8 +17,8 @@ import org.springframework.web.multipart.MultipartFile;
 import org.waldreg.board.board.management.BoardManager;
 import org.waldreg.board.board.management.BoardManager.BoardRequest;
 import org.waldreg.board.dto.BoardDto;
-import org.waldreg.board.board.exception.BoardDeletePermissionException;
-import org.waldreg.board.board.exception.BoardModifyPermissionException;
+import org.waldreg.board.exception.BoardDeletePermissionException;
+import org.waldreg.board.exception.BoardModifyPermissionException;
 import org.waldreg.board.file.FileManager;
 import org.waldreg.character.aop.annotation.PermissionVerifying;
 import org.waldreg.character.aop.behavior.VerifyingFailBehavior;
