--- waldreg/service-layer/board/src/main/java/org/waldreg/board/board/exception/BoardDoesNotExistException.java
@@ -1,9 +1,0 @@
-package org.waldreg.board.board.exception;
-
-public class BoardDoesNotExistException extends RuntimeException{
-
-    public BoardDoesNotExistException(int boardId){
-        super("Board Does Not Exist\nBoardId: " + boardId);
-    }
-
-}
