--- waldreg/service-layer/board/src/main/java/org/waldreg/board/board/exception/CategoryDoesNotExistException.java
@@ -1,9 +1,0 @@
-package org.waldreg.board.board.exception;
-
-public class CategoryDoesNotExistException extends RuntimeException{
-
-    public CategoryDoesNotExistException(int categoryId){
-        super("Category Does Not Exist\nCategoryId:  " + categoryId);
-    }
-
-}
