+++ waldreg/service-layer/board/src/main/java/org/waldreg/board/board/management/DefaultBoardManager.java
@@ -7,10 +7,10 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.waldreg.board.board.exception.BoardDoesNotExistException;
-import org.waldreg.board.board.exception.CategoryDoesNotExistException;
-import org.waldreg.board.board.exception.FutureCannotGetException;
-import org.waldreg.board.board.exception.InvalidRangeException;
+import org.waldreg.board.exception.BoardDoesNotExistException;
+import org.waldreg.board.exception.CategoryDoesNotExistException;
+import org.waldreg.board.exception.FutureCannotGetException;
+import org.waldreg.board.exception.InvalidRangeException;
 import org.waldreg.board.board.file.FileInfoGettable;
 import org.waldreg.board.board.spi.BoardRepository;
 import org.waldreg.board.board.spi.CategoryPerformer;
