+++ waldreg/service-layer/board/src/main/java/org/waldreg/board/category/management/DefaultCategoryManager.java
@@ -1,8 +1,8 @@
 package org.waldreg.board.category.management;
 
 import java.util.List;
-import org.waldreg.board.category.exception.CategoryDoesNotExistException;
-import org.waldreg.board.category.exception.DuplicateCategoryNameException;
+import org.waldreg.board.exception.CategoryDoesNotExistException;
+import org.waldreg.board.exception.DuplicateCategoryNameException;
 import org.waldreg.board.category.spi.CategoryRepository;
 import org.waldreg.board.dto.CategoryDto;
