--- waldreg/service-layer/board/src/main/java/org/waldreg/board/comment/exception/InvalidRangeException.java
@@ -1,9 +1,0 @@
-package org.waldreg.board.comment.exception;
-
-public class InvalidRangeException extends RuntimeException{
-
-    public InvalidRangeException(int startIdx, int endIdx){
-        super("Invalid range\nstartIdx : " + startIdx + " endIdx : " + endIdx);
-    }
-
-}
