+++ waldreg/service-layer/board/src/main/java/org/waldreg/board/comment/management/DefaultCommentManager.java
@@ -2,10 +2,10 @@ package org.waldreg.board.comment.management;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.waldreg.board.comment.exception.BoardDoesNotExistException;
-import org.waldreg.board.comment.exception.CommentDoesNotExistException;
-import org.waldreg.board.comment.exception.ContentLengthOverThousandException;
-import org.waldreg.board.comment.exception.InvalidRangeException;
+import org.waldreg.board.exception.BoardDoesNotExistException;
+import org.waldreg.board.exception.CommentDoesNotExistException;
+import org.waldreg.board.exception.ContentLengthOverThousandException;
+import org.waldreg.board.exception.InvalidRangeException;
 import org.waldreg.board.comment.spi.CommentInBoardRepository;
 import org.waldreg.board.comment.spi.CommentRepository;
 import org.waldreg.board.dto.CommentDto;
