+++ waldreg/service-layer/board/src/main/java/org/waldreg/board/exception/BoardDeletePermissionException.java
@@ -1,4 +1,4 @@
-package org.waldreg.board.board.exception;
+package org.waldreg.board.exception;
 
 public class BoardDeletePermissionException extends RuntimeException{
