+++ waldreg/service-layer/board/src/main/java/org/waldreg/board/exception/CategoryDoesNotExistException.java
@@ -1,9 +1,9 @@
-package org.waldreg.board.category.exception;
+package org.waldreg.board.exception;
 
 public class CategoryDoesNotExistException extends RuntimeException{
 
     public CategoryDoesNotExistException(int categoryId){
-        super("Category does not exist CategoryId : " + categoryId);
+        super("Category Does Not Exist\nCategoryId:  " + categoryId);
     }
 
 }
