+++ waldreg/service-layer/board/src/main/java/org/waldreg/board/exception/CommentDoesNotExistException.java
@@ -1,4 +1,4 @@
-package org.waldreg.board.comment.exception;
+package org.waldreg.board.exception;
 
 public class CommentDoesNotExistException extends RuntimeException{
