+++ waldreg/service-layer/board/src/main/java/org/waldreg/board/exception/ContentLengthOverThousandException.java
@@ -1,4 +1,4 @@
-package org.waldreg.board.comment.exception;
+package org.waldreg.board.exception;
 
 public class ContentLengthOverThousandException extends RuntimeException{
