+++ waldreg/service-layer/board/src/main/java/org/waldreg/board/exception/DuplicateCategoryNameException.java
@@ -1,4 +1,4 @@
-package org.waldreg.board.category.exception;
+package org.waldreg.board.exception;
 
 public class DuplicateCategoryNameException extends RuntimeException{
