+++ waldreg/service-layer/board/src/main/java/org/waldreg/board/exception/FutureCannotGetException.java
@@ -1,4 +1,4 @@
-package org.waldreg.board.board.exception;
+package org.waldreg.board.exception;
 
 public class FutureCannotGetException extends RuntimeException{
