+++ waldreg/service-layer/board/src/main/java/org/waldreg/board/exception/ReactionTypeDoesNotExistException.java
@@ -1,4 +1,4 @@
-package org.waldreg.board.reaction.exception;
+package org.waldreg.board.exception;
 
 public class ReactionTypeDoesNotExistException extends RuntimeException{
