+++ waldreg/service-layer/board/src/test/java/org/waldreg/board/category/CategoryManagerTest.java
@@ -13,8 +13,8 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.waldreg.board.board.spi.UserRepository;
-import org.waldreg.board.category.exception.CategoryDoesNotExistException;
-import org.waldreg.board.category.exception.DuplicateCategoryNameException;
+import org.waldreg.board.exception.CategoryDoesNotExistException;
+import org.waldreg.board.exception.DuplicateCategoryNameException;
 import org.waldreg.board.category.management.CategoryManager;
 import org.waldreg.board.category.management.DefaultCategoryManager;
 import org.waldreg.board.category.spi.CategoryRepository;
