+++ waldreg/service-layer/board/src/test/java/org/waldreg/board/comment/CommentManagerTest.java
@@ -13,10 +13,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.waldreg.board.comment.exception.BoardDoesNotExistException;
-import org.waldreg.board.comment.exception.CommentDoesNotExistException;
-import org.waldreg.board.comment.exception.ContentLengthOverThousandException;
-import org.waldreg.board.comment.exception.InvalidRangeException;
+import org.waldreg.board.exception.BoardDoesNotExistException;
+import org.waldreg.board.exception.CommentDoesNotExistException;
+import org.waldreg.board.exception.ContentLengthOverThousandException;
+import org.waldreg.board.exception.InvalidRangeException;
 import org.waldreg.board.comment.management.CommentManager;
 import org.waldreg.board.comment.management.DefaultCommentManager;
 import org.waldreg.board.comment.spi.CommentInBoardRepository;
