+++ waldreg/service-layer/board/src/test/java/org/waldreg/board/reaction/ReactionManagerTest.java
@@ -19,8 +19,8 @@ import org.waldreg.board.dto.BoardServiceReactionType;
 import org.waldreg.board.dto.ReactionDto;
 import org.waldreg.board.dto.ReactionRequestDto;
 import org.waldreg.board.dto.UserDto;
-import org.waldreg.board.reaction.exception.BoardDoesNotExistException;
-import org.waldreg.board.reaction.exception.ReactionTypeDoesNotExistException;
+import org.waldreg.board.exception.BoardDoesNotExistException;
+import org.waldreg.board.exception.ReactionTypeDoesNotExistException;
 import org.waldreg.board.reaction.management.DefaultReactionManager;
 import org.waldreg.board.reaction.management.ReactionManager;
 import org.waldreg.board.reaction.spi.ReactionInBoardRepository;
